Added asset image instead of delete icon. Renamed dustbin named class.
